SLEMA-79 Cleaned the code related to well being assessment
